fix(go): convert Bits constants to uint in BitLayout literal (restore main-green)
The multi-language lint gate has been red at the Go step: zeta_id.go failed to
typecheck — the generated *Offset/*Width constants are typed `Bits` (uint32)
but BitField.Offset/Width are `uint`, and Go has no implicit numeric conversion
("cannot use VersionOffset (constant 123 of uint32 type Bits) as uint value").

BitField.Offset/Width must stay `uint` because they feed big.Int.Lsh/Rsh, which
take uint — so convert the constants at the literal (the hand-written consumer),
not the generated .gen.go. Verified with the CI toolchain: go build / go vet /
gofmt clean, go test ./... green (incl. the cross-language oracle), and
golangci-lint v2.12.2 (CI version) reports 0 issues.

Restores the multi-language lint gate to green (the markdownlint half was #8113).

Agency-Signature-Version: 1
Agent: Otto
Agent-Runtime: Claude Code
Agent-Model: claude-opus-4-8
Credential-Identity: AceHack
Credential-Mode: shared
Human-Review: explicit
Human-Review-Evidence: chat
Action-Mode: human-directed
Task: none

+++ src/Core.Go/zeta_id/zeta_id.go
@@ -160,16 +160,19 @@ var BitLayout = struct {
 	Location   BitField
 	Randomness BitField
 }{
-	Version:    BitField{Offset: VersionOffset, Width: VersionWidth},
-	Timestamp:  BitField{Offset: TimestampOffset, Width: TimestampWidth},
-	Chromosome: BitField{Offset: ChromosomeOffset, Width: ChromosomeWidth},
-	Category:   BitField{Offset: CategoryOffset, Width: CategoryWidth},
-	Firefly:    BitField{Offset: FireflyOffset, Width: FireflyWidth},
-	Authority:  BitField{Offset: AuthorityOffset, Width: AuthorityWidth},
-	Persona:    BitField{Offset: PersonaOffset, Width: PersonaWidth},
-	Momentum:   BitField{Offset: MomentumOffset, Width: MomentumWidth},
-	Location:   BitField{Offset: LocationOffset, Width: LocationWidth},
-	Randomness: BitField{Offset: RandomnessOffset, Width: RandomnessWidth},
+	// The generated *Offset/*Width constants are typed `Bits` (uint32);
+	// BitField.Offset/Width are `uint` (they feed big.Int.Lsh/Rsh, which take
+	// uint). Go has no implicit numeric conversion, so convert at the boundary.
+	Version:    BitField{Offset: uint(VersionOffset), Width: uint(VersionWidth)},
+	Timestamp:  BitField{Offset: uint(TimestampOffset), Width: uint(TimestampWidth)},
+	Chromosome: BitField{Offset: uint(ChromosomeOffset), Width: uint(ChromosomeWidth)},
+	Category:   BitField{Offset: uint(CategoryOffset), Width: uint(CategoryWidth)},
+	Firefly:    BitField{Offset: uint(FireflyOffset), Width: uint(FireflyWidth)},
+	Authority:  BitField{Offset: uint(AuthorityOffset), Width: uint(AuthorityWidth)},
+	Persona:    BitField{Offset: uint(PersonaOffset), Width: uint(PersonaWidth)},
+	Momentum:   BitField{Offset: uint(MomentumOffset), Width: uint(MomentumWidth)},
+	Location:   BitField{Offset: uint(LocationOffset), Width: uint(LocationWidth)},
+	Randomness: BitField{Offset: uint(RandomnessOffset), Width: uint(RandomnessWidth)},
 }
 
 func setBits(target *big.Int, field BitField, val uint64) {
